+++ src/AdapterAmazonS3.ts
@@ -13,53 +13,56 @@ import {
   HeadBucketCommand,
   HeadObjectCommand,
   ListBucketsCommand,
+  ListObjectVersionsCommand,
   // ListObjectVersionsCommand,
   ListObjectsCommand,
   PutObjectCommand,
   S3Client,
 } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
-import { ConfigAmazonS3, AdapterConfig, StorageType, S3Compatible } from "./types";
+import {
+  ConfigAmazonS3,
+  AdapterConfig,
+  StorageType,
+  S3Compatible,
+  ResultObjectStream,
+  ResultObject,
+  ResultObjectBuckets,
+  FileBufferParams,
+  FilePathParams,
+  FileStreamParams,
+  ResultObjectFiles,
+  ResultObjectNumber,
+  ResultObjectBoolean,
+} from "./types";
 import { parseUrl } from "./util";
 
 export class AdapterAmazonS3 extends AbstractAdapter {
-  protected type = StorageType.S3;
-  private storage: S3Client;
-  private bucketNames: string[] = [];
-  private region: string = "";
-  private s3Compatible: S3Compatible = S3Compatible.Amazon;
+  protected _type = StorageType.S3;
   protected conf: ConfigAmazonS3;
+  private configError: string | null = null;
+  private storage: S3Client;
+  private s3Compatible: S3Compatible = S3Compatible.Amazon;
 
   constructor(config: string | AdapterConfig) {
     super();
     this.conf = this.parseConfig(config as ConfigAmazonS3);
-    if (typeof this.conf.bucketName !== "undefined" && this.conf.bucketName !== "") {
-      const msg = this.validateName(this.conf.bucketName);
-      if (msg !== null) {
-        throw new Error(msg);
-      }
-      this.bucketName = this.conf.bucketName;
-    }
 
+    // handle small differences in supported S3 compatible storages
     if (typeof (this.conf as ConfigAmazonS3).region === "undefined") {
       if (this.s3Compatible === S3Compatible.R2) {
         this.conf.region = "auto";
-        this.region = this.conf.region;
       } else if (this.s3Compatible === S3Compatible.Backblaze) {
         let ep = this.conf.endpoint;
         ep = ep.substring(ep.indexOf("s3.") + 3);
         this.conf.region = ep.substring(0, ep.indexOf("."));
-        // console.log(this.config.region);
-        this.region = this.conf.region;
       }
-    } else {
-      this.region = (this.conf as ConfigAmazonS3).region;
     }
     if (typeof this.conf.endpoint === "undefined") {
-      this.storage = new S3Client({ region: this.region });
+      this.storage = new S3Client({ region: this.conf.region });
     } else {
       this.storage = new S3Client({
-        region: this.region,
+        region: this.conf.region,
         endpoint: this.conf.endpoint,
         credentials: {
           accessKeyId: this.conf.accessKeyId,
@@ -69,27 +72,14 @@ export class AdapterAmazonS3 extends AbstractAdapter {
     }
   }
 
-  async init(): Promise<boolean> {
-    if (this.initialized) {
-      return Promise.resolve(true);
-    }
-    if (this.bucketName) {
-      await this.createBucket(this.bucketName)
-        .then((_data) => {
-          this.bucketNames.push(this.bucketName);
-        })
-        .catch((message: string) => {
-          return Promise.reject(message);
-        });
-    }
-    // no further initialization required
-    this.initialized = true;
-    return Promise.resolve(true);
-  }
-
-  private parseConfig(config: string | ConfigAmazonS3): ConfigAmazonS3 {
+  private parseConfig(config: string | ConfigAmazonS3): ConfigAmazonS3 | null {
     let cfg: ConfigAmazonS3;
     if (typeof config === "string") {
+      const { value, error } = parseUrl(config);
+      if (error) {
+        this.configError = error;
+        return null;
+      }
       const {
         type,
         part1: accessKeyId,
@@ -97,7 +87,7 @@ export class AdapterAmazonS3 extends AbstractAdapter {
         part3: region,
         bucketName,
         queryString,
-      } = parseUrl(config);
+      } = value;
       cfg = {
         type,
         accessKeyId,
@@ -115,9 +105,9 @@ export class AdapterAmazonS3 extends AbstractAdapter {
     }
 
     if (!cfg.accessKeyId || !cfg.secretAccessKey) {
-      throw new Error(
-        "You must specify a value for both 'applicationKeyId' and  'applicationKey' for storage type 's3'"
-      );
+      this.configError =
+        "You must specify a value for both 'applicationKeyId' and  'applicationKey' for storage type 's3'";
+      return null;
     }
 
     if (typeof cfg.endpoint !== "undefined") {
@@ -128,49 +118,58 @@ export class AdapterAmazonS3 extends AbstractAdapter {
       }
     }
     if (!cfg.region && this.s3Compatible === S3Compatible.Amazon) {
-      throw new Error("You must specify a default region for storage type 's3'");
+      this.configError = "You must specify a default region for storage type 's3'";
+      return null;
     }
 
     return cfg;
   }
 
   async getFileAsStream(
+    bucketName: string,
     fileName: string,
     options: { start?: number; end?: number } = { start: 0 }
-  ): Promise<Readable> {
+  ): Promise<ResultObjectStream> {
+    if (this.configError !== null) {
+      return { error: this.configError, value: null };
+    }
+
     const params = {
-      Bucket: this.bucketName,
+      Bucket: bucketName,
       Key: fileName,
       Range: `bytes=${options.start}-${options.end || ""}`,
     };
 
     const command = new GetObjectCommand(params);
-    const response = await this.storage.send(command);
-    return response.Body as Readable;
+    try {
+      const response = await this.storage.send(command);
+      return { value: response.Body as Readable, error: null };
+    } catch (e) {
+      return { value: null, error: e.code };
+    }
   }
 
-  async removeFile(fileName: string): Promise<string> {
+  async removeFile(bucketName: string, fileName: string): Promise<ResultObject> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
+    }
+
     const input = {
-      Bucket: this.bucketName,
+      Bucket: bucketName,
       Key: fileName,
     };
     const command = new DeleteObjectCommand(input);
-    const response = await this.storage.send(command);
-    // console.log(response);
-    return "file removed";
+    try {
+      const response = await this.storage.send(command);
+      return { value: "ok", error: null };
+    } catch (e) {
+      return { value: null, error: e.code };
+    }
   }
 
-  // util members
-
-  async createBucket(name: string, options: object = {}): Promise<string> {
-    // return Promise.reject("oops");
-    const msg = this.validateName(name);
-    if (msg !== null) {
-      return Promise.reject(msg);
-    }
-
-    if (this.bucketNames.findIndex((b) => b === name) !== -1) {
-      return "bucket exists";
+  async createBucket(name: string, options: object = {}): Promise<ResultObject> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
     }
 
     try {
@@ -180,10 +179,7 @@ export class AdapterAmazonS3 extends AbstractAdapter {
       const command = new HeadBucketCommand(input);
       const response = await this.storage.send(command);
       if (response.$metadata.httpStatusCode === 200) {
-        // console.log("response", response);
-        this.bucketNames.push(name);
-        this.bucketName = name;
-        return "bucket exists";
+        return { error: "bucket exists", value: null };
       }
     } catch (_e) {
       // this error simply means that the bucket doesn't exist yet
@@ -206,227 +202,228 @@ export class AdapterAmazonS3 extends AbstractAdapter {
       const response = await this.storage.send(command);
       // console.log("response", response);
       if (response.$metadata.httpStatusCode === 200) {
-        this.bucketNames.push(name);
-        this.bucketName = name;
-        return "bucket created";
+        return { value: "ok", error: null };
+      } else {
+        return {
+          error: `Error http status code ${response.$metadata.httpStatusCode}`,
+          value: null,
+        };
+      }
+    } catch (error) {
+      return { error, value: null };
+    }
+  }
+
+  async clearBucket(name: string): Promise<ResultObject> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
+    }
+
+    try {
+      const input1 = {
+        Bucket: name,
+        MaxKeys: 1000,
+      };
+      const command = new ListObjectVersionsCommand(input1);
+      const { Versions } = await this.storage.send(command);
+      // console.log("Versions", Versions);
+      if (typeof Versions === "undefined") {
+        return { value: "bucket is empty", error: null };
+      }
+
+      try {
+        const input2 = {
+          Bucket: name,
+          Delete: {
+            Objects: Versions.map((value) => ({
+              Key: value.Key,
+              VersionId: value.VersionId,
+            })),
+            Quiet: false,
+          },
+        };
+        const command2 = new DeleteObjectsCommand(input2);
+        await this.storage.send(command2);
+        return { value: "ok", error: null };
+      } catch (e) {
+        return { value: null, error: e.code };
       }
     } catch (e) {
-      return Promise.reject(e.message);
+      return { value: null, error: e.code };
     }
   }
 
-  async selectBucket(name: string | null): Promise<string> {
-    // add check if bucket exists!
-    if (!name) {
-      this.bucketName = "";
-      return `bucket '${name}' deselected`;
-    }
-    await this.createBucket(name);
-    this.bucketName = name;
-    return `bucket '${name}' selected`;
-  }
-
-  async clearBucket(name?: string): Promise<string> {
-    const n = name || this.bucketName;
-
-    if (!n) {
-      throw new Error("no bucket selected");
+  async deleteBucket(name: string): Promise<ResultObject> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
     }
 
-    /*
-    const input1 = {
-      Bucket: n,
-      MaxKeys: 1000,
-    };
-    const command = new ListObjectVersionsCommand(input1);
-    const { Versions } = await this.storage.send(command);
-    // console.log("Versions", Versions);
-    if (typeof Versions === "undefined") {
-      return "bucket is empty";
-    }
-    const input2 = {
-      Bucket: n,
-      Delete: {
-        Objects: Versions.map((value) => ({
-          Key: value.Key,
-          VersionId: value.VersionId,
-        })),
-        Quiet: false,
-      },
-    };
-    const command2 = new DeleteObjectsCommand(input2);
-    const response = await this.storage.send(command2);
-    return "bucket cleared";
-    */
-
-    const input1 = {
-      Bucket: n,
-      MaxKeys: 1000,
-    };
-    const command1 = new ListObjectsCommand(input1);
-    const response1 = await this.storage.send(command1);
-    const Contents = response1.Contents;
-
-    if (!Contents || Contents.length === 0) {
-      return;
-    }
-    // console.log(Contents);
-    const input2 = {
-      Bucket: n,
-      Delete: {
-        Objects: Contents.map((value) => ({ Key: value.Key })),
-        Quiet: false,
-      },
-    };
-    const command2 = new DeleteObjectsCommand(input2);
-    const response = await this.storage.send(command2);
-    // console.log(response);
-    return "bucket cleared";
-  }
-
-  async deleteBucket(name?: string): Promise<string> {
-    const n = name || this.bucketName;
-    // console.log("deleteBucket", n);
-
-    if (n === "") {
-      throw new Error("deleteBucket: no bucket selected");
-    }
     try {
       const input = {
-        Bucket: n,
+        Bucket: name,
       };
       const command = new DeleteBucketCommand(input);
       const response = await this.storage.send(command);
       // console.log(response);
-
-      if (n === this.bucketName) {
-        this.bucketName = "";
-      }
-      // console.log("selected bucket", this.bucketName);
-      this.bucketNames = this.bucketNames.filter((b) => b !== n);
-
-      return "bucket deleted";
+      return { value: "ok", error: null };
     } catch (e) {
       if (e.code === "NoSuchBucket") {
-        throw new Error("bucket not found");
+        return { value: "bucket not found", error: null };
       }
-      throw e;
+      return { value: "bucket not found", error: e.code };
     }
   }
 
-  async listBuckets(): Promise<string[]> {
+  async listBuckets(): Promise<ResultObjectBuckets> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
+    }
+
     const input = {};
     const command = new ListBucketsCommand(input);
     return this.storage
       .send(command)
       .then((response) => {
-        this.bucketNames = response.Buckets?.map((d) => d?.Name);
-        return this.bucketNames;
+        const bucketNames = response.Buckets?.map((d) => d?.Name);
+        return { value: bucketNames, error: null };
       })
       .catch((e) => {
-        console.log("[ERROR listBuckets]", e);
-        return [];
+        return { value: null, error: e.code };
       });
   }
 
-  protected async store(buffer: Buffer, targetPath: string, options: object): Promise<string>;
-  protected async store(stream: Readable, targetPath: string, options: object): Promise<string>;
-  protected async store(origPath: string, targetPath: string, options: object): Promise<string>;
-  protected async store(
-    arg: string | Buffer | Readable,
-    targetPath: string,
-    options: object
-  ): Promise<string> {
-    if (!this.bucketName) {
-      throw new Error("no bucket selected");
+  public async addFile(
+    params: FilePathParams | FileBufferParams | FileStreamParams
+  ): Promise<ResultObject> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
     }
+
+    let { options } = params;
     if (typeof options !== "object") {
       options = {};
     }
-    await this.createBucket(this.bucketName);
 
-    const input = {
-      ...options,
-      Bucket: this.bucketName,
-      Key: targetPath,
-      Body: arg,
-    };
-
-    if (typeof arg === "string") {
-      if (!fs.existsSync(arg)) {
-        throw new Error(`File with given path: ${arg}, was not found`);
+    let fileData: Readable | Buffer;
+    if (typeof (params as FilePathParams).origPath !== "undefined") {
+      const f = (params as FilePathParams).origPath;
+      if (!fs.existsSync(f)) {
+        return { value: null, error: `File with given path: ${f}, was not found` };
       }
-      input.Body = fs.createReadStream(arg);
+      fileData = fs.createReadStream(f);
+    } else if (typeof (params as FileBufferParams).buffer !== "undefined") {
+      fileData = (params as FileBufferParams).buffer;
+    } else if (typeof (params as FileStreamParams).stream !== "undefined") {
+      fileData = (params as FileStreamParams).stream;
     }
 
-    const command = new PutObjectCommand(input);
-    const response = await this.storage.send(command);
-
-    if (this.region !== "") {
+    try {
       const input = {
-        Bucket: this.bucketName,
+        Bucket: params.bucketName,
+        Key: params.targetPath,
+        Body: fileData,
+        ...options,
       };
-      const command = new GetBucketLocationCommand(input);
+      const command = new PutObjectCommand(input);
       const response = await this.storage.send(command);
-      this.region = response.LocationConstraint;
+      return this.getFileAsURL(params.bucketName, params.targetPath);
+    } catch (e) {
+      return { value: null, error: e.code };
     }
-    // return `https://${this.bucketName}.s3.${this.region}.amazonaws.com/${targetPath}`;
-    return await getSignedUrl(
+  }
+
+  async getFileAsURL(bucketName: string, fileName: string): Promise<ResultObject> {
+    return getSignedUrl(
       this.storage,
-      new GetObjectCommand({ Bucket: this.bucketName, Key: targetPath })
+      new GetObjectCommand({
+        Bucket: bucketName,
+        Key: fileName,
+      })
       // { expiresIn: 3600 }
-    );
+    )
+      .then((url: string) => {
+        return { value: url, error: null };
+      })
+      .catch((e) => {
+        return { value: null, error: e.code };
+      });
   }
 
-  async listFiles(maxFiles: number = 1000): Promise<[string, number][]> {
-    if (!this.bucketName) {
-      throw new Error("no bucket selected");
+  async listFiles(bucketName: string, maxFiles: number = 1000): Promise<ResultObjectFiles> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
+    }
+    try {
+      const input = {
+        Bucket: bucketName,
+        MaxKeys: maxFiles,
+      };
+      const command = new ListObjectsCommand(input);
+      const response = await this.storage.send(command);
+      const { Contents } = response;
+      if (!Contents) {
+        return { value: [], error: null };
+      }
+      return { value: Contents.map((o) => [o.Key, o.Size]), error: null };
+    } catch (e) {
+      return { value: null, error: e.code };
+    }
+  }
+
+  async sizeOf(bucketName: string, fileName: string): Promise<ResultObjectNumber> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
+    }
+
+    try {
+      const input = {
+        Bucket: bucketName,
+        Key: fileName,
+      };
+      const command = new HeadObjectCommand(input);
+      const response = await this.storage.send(command);
+      return { value: response.ContentLength, error: null };
+    } catch (e) {
+      return { value: null, error: e.code };
+    }
+  }
+
+  async bucketExists(bucketName: string): Promise<ResultObjectBoolean> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
     }
 
     const input = {
-      Bucket: this.bucketName,
-      MaxKeys: maxFiles,
+      Bucket: bucketName,
     };
-    const command = new ListObjectsCommand(input);
-    const response = await this.storage.send(command);
-    const { Contents } = response;
-    if (!Contents) {
-      return [];
-    }
-    return Contents.map((o) => [o.Key, o.Size]) as [string, number][];
+    const command = new HeadBucketCommand(input);
+    return this.storage
+      .send(command)
+      .then(() => {
+        return { value: true, error: null };
+      })
+      .catch(() => {
+        return { value: false, error: null };
+      });
   }
 
-  async sizeOf(name: string): Promise<number> {
-    if (!this.bucketName) {
-      throw new Error("no bucket selected");
+  async fileExists(bucketName: string, fileName: string): Promise<ResultObjectBoolean> {
+    if (this.configError !== null) {
+      return { value: null, error: this.configError };
     }
 
     const input = {
-      Bucket: this.bucketName,
-      Key: name,
-    };
-    const command = new HeadObjectCommand(input);
-    const response = await this.storage.send(command);
-    return response.ContentLength;
-  }
-
-  async fileExists(name: string): Promise<boolean> {
-    if (!this.bucketName) {
-      throw new Error("no bucket selected");
-    }
-
-    const input = {
-      Bucket: this.bucketName,
-      Key: name,
+      Bucket: bucketName,
+      Key: fileName,
     };
     const command = new HeadObjectCommand(input);
     return this.storage
       .send(command)
       .then(() => {
-        return true;
+        return { value: true, error: null };
       })
       .catch(() => {
-        return false;
+        return { value: false, error: null };
       });
   }
 }