+++ tests/test-config-local.jasmine.ts
@@ -5,13 +5,6 @@ import { rimraf } from "rimraf";
 import { Storage } from "../src/Storage";
 import { StorageType, ConfigLocal } from "../src/types";
 
-// describe("test jasmine", () => {
-//   it("weird", () => {
-//     expect("false").toBeTruthy();
-//     expect("0").toBeTruthy();
-//   });
-// });
-
 describe(`testing local urls`, () => {
   beforeAll(async () => {
     // override Linux's umask 0o002
@@ -20,37 +13,36 @@ describe(`testing local urls`, () => {
 
   it("[0]", async () => {
     const storage = new Storage("local://tests/tmp/the-buck?param=value");
-    await storage.init();
-    expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.type).toBe(StorageType.LOCAL);
+    expect((storage.config as ConfigLocal).directory).toBe("the-tests/tmp/");
+    expect(storage.config.bucketName).toBe("the-buck");
   });
   it("[0a]", async () => {
     const storage = new Storage("local://the-buck?param=value");
-    await storage.init();
-    expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.type).toBe(StorageType.LOCAL);
+    expect(storage.config.bucketName).toBe("the-buck");
+    expect(storage.config["param"]).toBe("value");
   });
 
   it("[1]", async () => {
     const storage = new Storage("local://tests/tmp");
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.config.bucketName).toBe("tmp");
   });
 
   it("[2] store in folder where process runs", async () => {
     const storage = new Storage(`local://${process.cwd()}/the-buck`);
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect((storage.config as ConfigLocal).directory).toBe(process.cwd());
   });
 
   it("[3]", async () => {
     const storage = new Storage({
       type: StorageType.LOCAL,
       directory: "tests/tmp/the-buck",
+      // versioning: true,
     });
-    await storage.init();
     expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("");
+    expect(storage.config.bucketName).toBe(undefined);
+    // expect(storage.config.versioning).toBe(undefined);
   });
 
   it("[4]", async () => {
@@ -59,9 +51,8 @@ describe(`testing local urls`, () => {
       directory: "tests/tmp",
       bucketName: "the-buck",
     });
-    await storage.init();
     expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.config.bucketName).toBe("the-buck");
   });
 
   it("[5] numeric values in options stay numeric and keep their radix (8)", async () => {
@@ -70,7 +61,6 @@ describe(`testing local urls`, () => {
       directory: "tests/tmp",
       mode: 0o777,
     });
-    await storage.init();
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(0o777);
   });
 
@@ -80,14 +70,12 @@ describe(`testing local urls`, () => {
       directory: "tests/tmp",
       mode: 511,
     });
-    await storage.init();
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(511);
   });
 
   it("[6] string values in options stay string values (will be converted when used in code when necessary)", async () => {
     const storage = new Storage("local://tests/tmp?mode=0o777");
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.config.bucketName).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("0o777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
     expect(mode.toString(8)).toBe("40777");
@@ -96,8 +84,7 @@ describe(`testing local urls`, () => {
 
   it("[6a]", async () => {
     const storage = new Storage("local://tests/tmp?mode=777");
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.config.bucketName).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
     expect(mode.toString(8)).toBe("40777");
@@ -111,8 +98,7 @@ describe(`testing local urls`, () => {
       bucketName: "the-buck",
       mode: 0o777,
     });
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.config.bucketName).toBe("the-buck");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(511);
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp", "the-buck")))
       .mode;
@@ -127,8 +113,7 @@ describe(`testing local urls`, () => {
       bucketName: "the-buck",
       mode: 511,
     });
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.config.bucketName).toBe("the-buck");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(511);
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp", "the-buck")))
       .mode;
@@ -138,8 +123,7 @@ describe(`testing local urls`, () => {
 
   it("[6e]", async () => {
     const storage = new Storage("local://tests/tmp?mode=0o777");
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.config.bucketName).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("0o777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
     expect(mode.toString(8)).toBe("40777");
@@ -154,8 +138,7 @@ describe(`testing local urls`, () => {
       bucketName: "tmp",
       mode: "755", // this is an error! the parseMode function will return the default value 0o777
     });
-    await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.config.bucketName).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("755");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
     expect(mode.toString(8)).toBe("40777");